Refactor into two packages
