+++ packages/mkc-node/tsconfig.json
@@ -1,0 +1,19 @@
+{
+    "compilerOptions": {
+        "noImplicitAny": true,
+        "noImplicitReturns": true,
+        "module": "commonjs",
+        "target": "es6",
+        "outDir": "built",
+        "lib": ["es6", "DOM"],
+        "sourceMap": true,
+        "types": ["node"],
+        "rootDir": "src",
+        "declaration": true
+    },
+    "exclude": ["node_modules", ".vscode-test"],
+    "include": [
+        "src/*.ts"
+    ]
+}
+
